+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `ItemList` type under `Intangible` along with unit tests and documentation.
+- Made `BreadcrumbList` extend `ItemList`.
+
 ## 0.4.0 - 2026-09-12
 
 ### Changed